Updated copyright

+++ YCML/Data Frame/NSIndexSet+Sampling.h
@@ -2,14 +2,28 @@
 //  NSIndexSet+Sampling.h
 //  YCML
 //
-//  Created by Ioannis Chatzikonstantinou on 1/7/15.
-//  Copyright (c) 2015 Yannis Chatzikonstantinou. All rights reserved.
+//  Created by Ioannis (Yannis) Chatzikonstantinou on 1/7/15.
+//  Copyright © 2015 Ioannis (Yannis) Chatzikonstantinou. All rights reserved.
 //
+// This file is part of YCML.
+//
+// YCML is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// YCML is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with YCML.  If not, see <http://www.gnu.org/licenses/>.
 
 #import <Foundation/Foundation.h>
 
 @interface NSIndexSet (Sampling)
 
-+ (instancetype)indexesForSampling:(int)samples inRange:(NSRange)range replacement:(BOOL)replacement;
++ (instancetype)indexesForSampling:(NSUInteger)samples inRange:(NSRange)range replacement:(BOOL)replacement;
 
 @end
